+++ CHANGELOG.md
@@ -6,6 +6,7 @@ All notable changes to this project will be documented in this file.
 
 ### Added
 
+- Automatic plot title from log file name
 - Warn when trying to plot structured data
 
 ## [0.3.0] - 2023/09/04